Move element base class methods into element interface as default methods to reduce code duplication
