+++ accumulo/src/main/java/org/vertexium/accumulo/AccumuloElement.java
@@ -6,11 +6,18 @@ import org.apache.hadoop.io.Text;
 import org.vertexium.*;
 import org.vertexium.accumulo.iterator.ElementIterator;
 import org.vertexium.mutation.*;
+import org.vertexium.property.MutableProperty;
+import org.vertexium.property.PropertyValue;
 import org.vertexium.query.ExtendedDataQueryableIterable;
 import org.vertexium.query.QueryableIterable;
+import org.vertexium.util.PropertyCollection;
 
 import java.io.Serializable;
 import java.util.EnumSet;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.Set;
+import java.util.concurrent.ConcurrentSkipListSet;
 
 public abstract class AccumuloElement extends ElementBase implements Serializable, HasTimestamp {
     private static final long serialVersionUID = 1L;
@@ -31,6 +38,19 @@ public abstract class AccumuloElement extends ElementBase implements Serializabl
     public static final Text METADATA_COLUMN_FAMILY = ElementIterator.METADATA_COLUMN_FAMILY;
     public static final Text METADATA_COLUMN_QUALIFIER = ElementIterator.METADATA_COLUMN_QUALIFIER;
 
+    private final Graph graph;
+    private final String id;
+    private Visibility visibility;
+    private final long timestamp;
+    private final EnumSet<FetchHint> fetchHints;
+    private Set<Visibility> hiddenVisibilities = new HashSet<>();
+
+    private final PropertyCollection properties;
+    private final ImmutableSet<String> extendedDataTableNames;
+    private ConcurrentSkipListSet<PropertyDeleteMutation> propertyDeleteMutations;
+    private ConcurrentSkipListSet<PropertySoftDeleteMutation> propertySoftDeleteMutations;
+    private final Authorizations authorizations;
+
     protected AccumuloElement(
             Graph graph,
             String id,
@@ -44,65 +64,36 @@ public abstract class AccumuloElement extends ElementBase implements Serializabl
             EnumSet<FetchHint> fetchHints,
             Authorizations authorizations
     ) {
-        super(
-                graph,
-                id,
-                visibility,
-                properties,
-                propertyDeleteMutations,
-                propertySoftDeleteMutations,
-                hiddenVisibilities,
-                extendedDataTableNames,
-                timestamp,
-                fetchHints,
-                authorizations
-        );
-    }
-
-    @Override
-    public void deleteProperty(String key, String name, Authorizations authorizations) {
-        Property property = super.removePropertyInternal(key, name);
-        if (property != null) {
-            getGraph().deleteProperty(this, property, authorizations);
+        this.graph = graph;
+        this.id = id;
+        this.visibility = visibility;
+        this.timestamp = timestamp;
+        this.fetchHints = fetchHints;
+        this.properties = new PropertyCollection();
+        this.extendedDataTableNames = extendedDataTableNames;
+        this.authorizations = authorizations;
+        if (hiddenVisibilities != null) {
+            for (Visibility v : hiddenVisibilities) {
+                this.hiddenVisibilities.add(v);
+            }
         }
+        updatePropertiesInternal(properties, propertyDeleteMutations, propertySoftDeleteMutations);
     }
 
     @Override
     public void deleteProperty(String key, String name, Visibility visibility, Authorizations authorizations) {
-        Property property = super.removePropertyInternal(key, name, visibility);
+        Property property = getProperty(key, name, visibility);
         if (property != null) {
+            this.properties.removeProperty(property);
             getGraph().deleteProperty(this, property, authorizations);
-        }
-    }
-
-    @Override
-    public void softDeleteProperty(String key, String name, Authorizations authorizations) {
-        Property property = super.softDeletePropertyInternal(key, name);
-        if (property != null) {
-            getGraph().softDeleteProperty(this, property, authorizations);
         }
     }
 
     @Override
     public void softDeleteProperty(String key, String name, Visibility visibility, Authorizations authorizations) {
-        Property property = super.softDeletePropertyInternal(key, name, visibility);
+        Property property = getProperty(key, name, visibility);
         if (property != null) {
-            getGraph().softDeleteProperty(this, property, authorizations);
-        }
-    }
-
-    @Override
-    public void deleteProperties(String name, Authorizations authorizations) {
-        Iterable<Property> properties = super.removePropertyInternal(name);
-        for (Property property : properties) {
-            getGraph().deleteProperty(this, property, authorizations);
-        }
-    }
-
-    @Override
-    public void softDeleteProperties(String name, Authorizations authorizations) {
-        Iterable<Property> properties = super.removePropertyInternal(name);
-        for (Property property : properties) {
+            this.properties.removeProperty(property);
             getGraph().softDeleteProperty(this, property, authorizations);
         }
     }
@@ -119,12 +110,7 @@ public abstract class AccumuloElement extends ElementBase implements Serializabl
 
     @Override
     public AccumuloGraph getGraph() {
-        return (AccumuloGraph) super.getGraph();
-    }
-
-    @Override
-    protected void setVisibility(Visibility visibility) {
-        super.setVisibility(visibility);
+        return (AccumuloGraph) graph;
     }
 
     protected <TElement extends Element> void saveExistingElementMutation(ExistingElementMutationImpl<TElement> mutation, Authorizations authorizations) {
@@ -198,5 +184,200 @@ public abstract class AccumuloElement extends ElementBase implements Serializabl
                 tableName,
                 getGraph().getExtendedData(ElementType.getTypeFromElement(this), getId(), tableName, getAuthorizations())
         );
+    }
+
+    @Override
+    public Property getProperty(String key, String name) {
+        return getProperty(key, name, null);
+    }
+
+    @Override
+    public Property getProperty(String name) {
+        if (ID_PROPERTY_NAME.equals(name)) {
+            return getIdProperty();
+        } else if (Edge.LABEL_PROPERTY_NAME.equals(name) && this instanceof Edge) {
+            return getEdgeLabelProperty();
+        }
+        Iterator<Property> propertiesWithName = getProperties(name).iterator();
+        if (propertiesWithName.hasNext()) {
+            return propertiesWithName.next();
+        }
+        return null;
+    }
+
+    @Override
+    public Object getPropertyValue(String name, int index) {
+        if (ID_PROPERTY_NAME.equals(name)) {
+            return getIdProperty();
+        } else if (Edge.LABEL_PROPERTY_NAME.equals(name) && this instanceof Edge) {
+            return getEdgeLabelProperty();
+        }
+        Property property = this.properties.getProperty(name, index);
+        if (property == null) {
+            return null;
+        }
+        return property.getValue();
+    }
+
+    @Override
+    public Object getPropertyValue(String key, String name, int index) {
+        if (ID_PROPERTY_NAME.equals(name)) {
+            return getIdProperty();
+        } else if (Edge.LABEL_PROPERTY_NAME.equals(name) && this instanceof Edge) {
+            return getEdgeLabelProperty();
+        }
+        Property property = this.properties.getProperty(key, name, index);
+        if (property == null) {
+            return null;
+        }
+        return property.getValue();
+    }
+
+    @Override
+    public String getId() {
+        return this.id;
+    }
+
+    @Override
+    public Visibility getVisibility() {
+        return this.visibility;
+    }
+
+    @Override
+    public long getTimestamp() {
+        return timestamp;
+    }
+
+    protected void setVisibility(Visibility visibility) {
+        this.visibility = visibility;
+    }
+
+    @Override
+    public Iterable<Property> getProperties() {
+        return this.properties.getProperties();
+    }
+
+    public Iterable<PropertyDeleteMutation> getPropertyDeleteMutations() {
+        return this.propertyDeleteMutations;
+    }
+
+    public Iterable<PropertySoftDeleteMutation> getPropertySoftDeleteMutations() {
+        return this.propertySoftDeleteMutations;
+    }
+
+    @Override
+    public Iterable<Property> getProperties(final String key, final String name) {
+        if (ID_PROPERTY_NAME.equals(name) || (Edge.LABEL_PROPERTY_NAME.equals(name) && this instanceof Edge)) {
+            return getProperties(name);
+        }
+        return this.properties.getProperties(key, name);
+    }
+
+    // this method differs setProperties in that it only updates the in memory representation of the properties
+    protected void updatePropertiesInternal(
+            Iterable<Property> properties,
+            Iterable<PropertyDeleteMutation> propertyDeleteMutations,
+            Iterable<PropertySoftDeleteMutation> propertySoftDeleteMutations
+    ) {
+        if (propertyDeleteMutations != null) {
+            this.propertyDeleteMutations = new ConcurrentSkipListSet<>();
+            for (PropertyDeleteMutation propertyDeleteMutation : propertyDeleteMutations) {
+                removePropertyInternal(
+                        propertyDeleteMutation.getKey(),
+                        propertyDeleteMutation.getName(),
+                        propertyDeleteMutation.getVisibility()
+                );
+                this.propertyDeleteMutations.add(propertyDeleteMutation);
+            }
+        }
+        if (propertySoftDeleteMutations != null) {
+            this.propertySoftDeleteMutations = new ConcurrentSkipListSet<>();
+            for (PropertySoftDeleteMutation propertySoftDeleteMutation : propertySoftDeleteMutations) {
+                removePropertyInternal(
+                        propertySoftDeleteMutation.getKey(),
+                        propertySoftDeleteMutation.getName(),
+                        propertySoftDeleteMutation.getVisibility()
+                );
+                this.propertySoftDeleteMutations.add(propertySoftDeleteMutation);
+            }
+        }
+
+        for (Property property : properties) {
+            addPropertyInternal(property);
+        }
+    }
+
+    protected void removePropertyInternal(String key, String name, Visibility visibility) {
+        Property property = getProperty(key, name, visibility);
+        if (property != null) {
+            this.properties.removeProperty(property);
+        }
+    }
+
+    protected void addPropertyInternal(Property property) {
+        if (property.getKey() == null) {
+            throw new IllegalArgumentException("key is required for property");
+        }
+        Object propertyValue = property.getValue();
+        if (propertyValue instanceof PropertyValue && !((PropertyValue) propertyValue).isStore()) {
+            return;
+        }
+        Property existingProperty = getProperty(property.getKey(), property.getName(), property.getVisibility());
+        if (existingProperty == null) {
+            this.properties.addProperty(property);
+        } else {
+            if (existingProperty instanceof MutableProperty) {
+                ((MutableProperty) existingProperty).update(property);
+            } else {
+                throw new VertexiumException("Could not update property of type: " + existingProperty.getClass().getName());
+            }
+        }
+    }
+
+    @Override
+    public int hashCode() {
+        return getId().hashCode();
+    }
+
+    @Override
+    public String toString() {
+        if (this instanceof Edge) {
+            Edge edge = (Edge) this;
+            return getId() + ":[" + edge.getVertexId(Direction.OUT) + "-" + edge.getLabel() + "->" + edge.getVertexId(Direction.IN) + "]";
+        }
+        return getId();
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj instanceof Element) {
+            Element objElem = (Element) obj;
+            return getId().equals(objElem.getId());
+        }
+        return super.equals(obj);
+    }
+
+    @Override
+    public Authorizations getAuthorizations() {
+        return authorizations;
+    }
+
+    public Iterable<Visibility> getHiddenVisibilities() {
+        return hiddenVisibilities;
+    }
+
+    @Override
+    public ImmutableSet<String> getExtendedDataTableNames() {
+        return extendedDataTableNames;
+    }
+
+    @Override
+    public EnumSet<FetchHint> getFetchHints() {
+        return fetchHints;
+    }
+
+    @Override
+    protected Iterable<Property> internalGetProperties(String name) {
+        return this.properties.getProperties(name);
     }
 }
